Move comment input to separate component
